Enhance authentication features and update dependencies
- Implement `GlobalExceptionHandler` for centralized error handling
- Add email-based rate limiting to login endpoints in `AuthenticationService`
- Track authentication provider (LOCAL, GOOGLE, APPLE) in `User` entity
- Add Apple token revocation logic in `SocialAuthService`
- Improve error handling in `UserService` and `RefreshTokenService` by replacing unsafe Optional usage
- Upgrade Spring Boot to 3.4.5 and refactor dependency versions in `pom.xml`
- Change JWT expiration type to long in `JwtUtils`

+++ src/main/java/com/dilem/framebackend/util/JwtUtils.java
@@ -4,7 +4,6 @@ import com.dilem.framebackend.model.User;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.UnsupportedJwtException;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
@@ -25,7 +24,7 @@ public class JwtUtils {
     private String jwtSecret;
 
     @Value("${application.security.jwt.access-token.expiration}")
-    private int jwtExpirationMs;
+    private long jwtExpirationMs;
 
     public String generateJwtToken(Authentication authentication) {
         User userPrincipal = (User) authentication.getPrincipal();
